webapp/requirements.txt
Flask==2.3.2
kafka-python==2.3.0